+++ docs/qml/ml_tools.md
@@ -113,7 +113,7 @@ import matplotlib.pyplot as plt
 
 from qadence import Parameter, QuantumCircuit, Z
 from qadence import hamiltonian_factory, hea, feature_map, chain
-from qadence.models import QNN
+from qadence import QNN
 from qadence.ml_tools import  TrainConfig, train_with_grad, to_dataloader
 
 DEVICE = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
@@ -173,7 +173,7 @@ import torch
 from itertools import count
 from qadence.constructors import hamiltonian_factory, hea, feature_map
 from qadence import chain, Parameter, QuantumCircuit, Z
-from qadence.models import QNN
+from qadence import QNN
 from qadence.ml_tools import train_with_grad, TrainConfig
 
 n_qubits = 2
