+++ qadence/draw/utils.py
@@ -10,6 +10,7 @@ from typing import Any
 
 import sympy
 
+from qadence import QuantumModel
 from qadence.blocks import (
     AbstractBlock,
     AddBlock,
@@ -23,7 +24,6 @@ from qadence.blocks import (
 )
 from qadence.blocks.analog import ConstantAnalogRotation, InteractionBlock
 from qadence.circuit import QuantumCircuit
-from qadence.models import QuantumModel
 from qadence.operations import RX, RY, RZ, SWAP, HamEvo, I
 from qadence.transpile.block import fill_identities
 from qadence.utils import format_parameter
