--- qadence/models/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from .qnn import QNN
-from .quantum_model import QuantumModel
-
-# Modules to be automatically added to the qadence namespace
-__all__ = ["QNN", "QuantumModel"]
