+++ qadence/serialization.py
@@ -10,12 +10,11 @@ import torch
 from sympy import *
 from sympy import Basic, Expr, srepr
 
-from qadence import QuantumCircuit, operations
+from qadence import QNN, QuantumCircuit, QuantumModel, operations
 from qadence import blocks as qadenceblocks
 from qadence.blocks import AbstractBlock
 from qadence.blocks.utils import tag
 from qadence.logger import get_logger
-from qadence.models import QNN, QuantumModel
 from qadence.parameters import Parameter
 from qadence.register import Register
 from qadence.types import SerializationFormat
