+++ tests/qadence/test_error_models.py
@@ -8,7 +8,7 @@ from numpy.random import rand
 from sympy import acos
 
 import qadence as qd
-from qadence import BackendName
+from qadence import BackendName, QuantumModel
 from qadence.blocks import (
     AbstractBlock,
     add,
@@ -18,7 +18,6 @@ from qadence.circuit import QuantumCircuit
 from qadence.constructors.hamiltonians import hamiltonian_factory
 from qadence.divergences import js_divergence
 from qadence.measurements.protocols import Measurements
-from qadence.models import QuantumModel
 from qadence.noise import Noise
 from qadence.noise.readout import WhiteNoise, bs_corruption, create_noise_matrix, sample_to_matrix
 from qadence.operations import (
